[WIP] Add Dockerfile
This commit adds a `Dockerfile` to the root of the project, accompanied
by a `docker-compose.yml` for simplified deployment. The `Dockerfile` is
agnostic to whether it will be the webserver, the consumer, or if it is
run for a one-off command (i.e. creation of a superuser, migration of
the database, document export, ...).

The containers entrypoint is the `scripts/docker-entrypoint.sh` script.
This script verifies that the required permissions are set, remaps the
default users and/or groups id if required and installs additional
languages if the user wishes to.

After initialization, it analyzes the command the user supplied:

  - If the command starts with a slash, it is expected that the user
    wants to execute a binary file and the command will be executed
    without further intervention. (Using `exec` to effectively replace
    the started shell-script and not have any reaping-issues.)

  - If the command does not start with a slash, the command will be
    passed directly to the `manage.py` script without further
    modification. (Again using `exec`.)

The default command is set to `--help`.

If the user wants to execute a command that is not meant for `manage.py`
but doesn't start with a slash, the Docker `--entrypoint` parameter can
be used to circumvent the mechanics of `docker-entrypoint.sh`.

Further information on user-remapping can be found in `docs/setup.rst`.